pyproject.toml
[project]
name = "intel-bot"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "openai>=1.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "python-dotenv>=1.0",
    "rich>=13.0",
    "wandb>=0.19",
    "wandb-workspaces>=0.3",
    "weave>=0.51",
    "crewai[tools]>=1.9",
    "markdown>=3.5",
    "playwright>=1.40",
]

[project.optional-dependencies]
dev = ["ruff", "mypy"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/intel_bot"]
